refactor: move useful code out of github.dispatch.actions.utils
