+++ openedx_webhooks/tasks/jira_work.py
@@ -22,6 +22,21 @@ def delete_jira_issue(issue_key):
     log_check_response(resp)
 
 
+def find_issues_for_pull_request(jira, pull_request_url):
+    """
+    Find corresponding JIRA issues for a given GitHub pull request.
+
+    Arguments:
+        jira (jira.JIRA): An authenticated JIRA API client session
+        pull_request_url (str)
+
+    Returns:
+        jira.client.ResultList[jira.Issue]
+    """
+    jql = 'project=OSPR AND cf[10904]="{}"'.format(pull_request_url)
+    return jira.search_issues(jql)
+
+
 def transition_jira_issue(issue_key, status_name):
     """
     Transition a Jira issue to a new status.
